+++ maplibre-gl-js/src/shaders/glsl/_prelude.vertex.glsl
@@ -177,6 +177,7 @@ layout(std140) uniform ProjectionUBO {
     highp mat4 u_projection_fallback_matrix;
     highp vec4 u_projection_tile_mercator_coords;
     highp vec4 u_projection_clipping_plane;
+    highp vec4 u_projection_origin;
     highp float u_projection_transition;
     highp int u_projection_clip_antimeridian;
 };