+++ maplibre-gl-js/src/shaders/glsl/symbol_text_and_icon.vertex.glsl
@@ -122,7 +122,7 @@ void main() {
     if (u_is_along_line || u_is_variable_anchor) {
         projected_pos = vec4(a_projected_pos.xy, ele, 1.0);
     } else if (u_pitch_with_map) {
-        projected_pos = u_label_plane_matrix * vec4(a_projected_pos.xy + u_translation, ele, 1.0);
+        projected_pos = u_label_plane_matrix * vec4(projectTileToPlane(a_projected_pos.xy + u_translation), ele, 1.0);
     } else {
         projected_pos = u_label_plane_matrix * projectTileWithElevation(a_projected_pos.xy + u_translation, ele);
     }
@@ -139,7 +139,7 @@ void main() {
 
     vec4 finalPos = u_coord_matrix * vec4(projected_pos.xy / projected_pos.w + rotation_matrix * (a_offset / 32.0 * fontScale) * projectionScaling, z, 1.0);
     if(u_pitch_with_map) {
-        finalPos = projectTileWithElevation(finalPos.xy, finalPos.z);
+        finalPos = projectPlanarTile(finalPos.xy, finalPos.z);
     }
     float gamma_scale = finalPos.w;
     gl_Position = finalPos;