+++ maplibre-gl-js/src/webgl/projection_uniform_buffer.ts
@@ -8,11 +8,13 @@ const layout = std140Layout([
     {name: 'u_projection_fallback_matrix', type: 'mat4'},
     {name: 'u_projection_tile_mercator_coords', type: 'vec4'},
     {name: 'u_projection_clipping_plane', type: 'vec4'},
+    {name: 'u_projection_origin', type: 'vec4'},
     {name: 'u_projection_transition', type: 'float'},
     {name: 'u_projection_clip_antimeridian', type: 'int'},
 ]);
 
 const offsets = layout.offsets;
+const defaultOrigin = [0, 0, 0, 0];
 
 /**
  * @internal
@@ -28,6 +30,7 @@ function updateProjectionUniformBuffer(buffer: UniformBuffer, projectionData: Pr
     f32.set(projectionData.fallbackMatrix, offsets.u_projection_fallback_matrix);
     f32.set(projectionData.tileMercatorCoords, offsets.u_projection_tile_mercator_coords);
     f32.set(projectionData.clippingPlane, offsets.u_projection_clipping_plane);
+    f32.set(projectionData.projectionOrigin ?? defaultOrigin, offsets.u_projection_origin);
     f32[offsets.u_projection_transition] = projectionData.projectionTransition;
     buffer.pendingWords[offsets.u_projection_clip_antimeridian] = projectionData.clipAntimeridian ? 1 : 0;
     buffer.upload();